Moved `Semaphore` and `Mutex` source to `Locking`
